> Changed the infogain code. Is this alright with you guys? Changed values from NormalDistriution to floats, so they look a bit better in the GUI. Set some default numbers to 1000, when not yet specified.
git-svn-id: https://svn.science.uu.nl/repos/sci.knobb101.SubgroupDiscovery@448 3a85a776-1598-11dc-ad04-adf9e166afdd

+++ SubDisc/src/nl/liacs/subdisc/NormalDistribution.java
@@ -90,9 +90,9 @@ public class NormalDistribution
 	public double getSigma() { return itsSigma; }
 	public double getVariance() { return itsSigma * itsSigma; }
 
-	public double getOnePercentSignificance() {return itsMu + 2.326348*itsSigma; }
-	public double getFivePercentSignificance() {return itsMu + 1.6448537*itsSigma; }
-	public double getTenPercentSignificance() {return itsMu + 1.2815517*itsSigma; }
+	public float getOnePercentSignificance() {return (float) (itsMu + 2.326348*itsSigma); }
+	public float getFivePercentSignificance() {return (float) (itsMu + 1.6448537*itsSigma); }
+	public float getTenPercentSignificance() {return (float) (itsMu + 1.2815517*itsSigma); }
 
 	public void printStatistics()
 	{
